+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   zones remains.
 - Externally available application listeners will now open their security group to both ingress and
   egress for their specified port.
+- Tweaked API.getFunction to allow [route] and [method] parameters to be optional.  Also changed
+  function to throw if passed arguments that don't map to an actual function.
 
 ## 0.18.4 (5/14/2019)